Modify LspConnection inteface
