--- tools/lspClient/LspConnectionInterface.ts
@@ -1,8 +1,0 @@
-export interface LspConnection {
-    initialize(): Promise<void>;
-    sendRequest(method: string, params: any): Promise<any>;
-    sendNotification(method: string, params: any): Promise<void>;
-    onNotification(method: string, handler: (params: any) => void): void;
-    onRequest(method: string, handler: (params: any) => any): void;
-    shutdown(): Promise<void>;
-}
